--- a/src/Core.cs
+++ b/src/Core.cs
@@ -6,11 +6,11 @@ namespace D2SLib;
 
 public class Core
 {
-    private static TXT? _TXT = null;
-    public static TXT TXT
+    private static MetaData? _metaData = null;
+    public static MetaData MetaData
     {
-        get => _TXT ?? ResourceFilesTXT.Instance.TXT;
-        set => _TXT = value;
+        get => _metaData ?? ResourceFilesData.Instance.MetaData;
+        set => _metaData = value;
     }
 
     public static D2S ReadD2S(string path) => D2S.Read(File.ReadAllBytes(path));