… benchmarks/D2SLib_Local.Benchmark/BenchmarkConfig.cs
… benchmarks/D2SLib_Local.Benchmark/D2SLib_Local.Benchmark.csproj
… benchmarks/D2SLib_Local.Benchmark/LoadGame.cs
… benchmarks/D2SLib_Local.Benchmark/Program.cs
… benchmarks/D2SLib_Nuget.Benchmark/D2SLib_Nuget.Benchmark.csproj
… benchmarks/D2SLib_Nuget.Benchmark/Program.cs
… benchmarks/Results.txt
… src/IO/BitField.cs
… src/IO/IBitReader.cs
… src/IO/InternalBitArray.cs
… src/MetaData.cs
… src/Model/Data/DataColumn.cs
… src/Model/Data/DataFile.cs
… src/Model/Data/ItemStatCostData.cs
… src/Model/Data/ItemsData.cs
… src/Model/Huffman/HuffmanTree.cs
… src/Model/Huffman/Node.cs
… src/Model/Save/Appearances.cs
… src/Model/Save/Attributes.cs
… src/Model/Save/Corpses.cs
… src/Model/Save/D2I.cs
… src/Model/Save/D2S.cs
… src/Model/Save/Golem.cs
… src/Model/Save/Header.cs
… src/Model/Save/Items.cs
… src/Model/Save/Locations.cs
… src/Model/Save/Mercenary.cs
… src/Model/Save/NPCDialogs.cs
… src/Model/Save/Quests.cs
… src/Model/Save/Skills.cs