I hope this rather large PR out of the blue isn't too presumptuous, but here's why you might want to take a closer look:

Load: **229** times faster, **77%** less memory used.
Save: **249** times faster, **90%** less memory used.

Version 1.0.2:

|   Method |       Runtime |     Mean |   Error |   StdDev |   Median |    Gen 0 | Allocated |
|--------- |-------------- |---------:|--------:|---------:|---------:|---------:|----------:|
| LoadOnly |      .NET 6.0 | 168.7 ms | 3.35 ms |  7.36 ms | 169.5 ms |        - |      1 MB |
| SaveOnly |      .NET 6.0 | 304.1 ms | 8.64 ms | 25.49 ms | 316.7 ms | 500.0000 |      2 MB |
| LoadOnly | .NET Core 3.1 | 169.7 ms | 3.37 ms |  7.68 ms | 170.3 ms |        - |      1 MB |
| SaveOnly | .NET Core 3.1 | 304.2 ms | 8.02 ms | 23.15 ms | 316.4 ms | 500.0000 |      2 MB |

This PR:

|   Method |       Runtime |       Mean |   Error |  StdDev |   Gen 0 |   Gen 1 | Allocated |
|--------- |-------------- |-----------:|--------:|--------:|--------:|--------:|----------:|
| LoadOnly |      .NET 6.0 |   735.6 us | 4.74 us | 4.20 us | 50.7813 | 11.7188 |    237 KB |
| SaveOnly |      .NET 6.0 | 1,223.6 us | 7.57 us | 8.42 us | 44.9219 |       - |    215 KB |
| LoadOnly | .NET Core 3.1 |   880.1 us | 3.63 us | 3.22 us | 50.7813 | 11.7188 |    237 KB |
| SaveOnly | .NET Core 3.1 | 1,426.6 us | 8.43 us | 7.88 us | 44.9219 |       - |    215 KB |

A lot of save-game editors are sketchy-seeming binary downloads, and non-programmers just have to trust that the editor isn't installing malware or mining bitcoin on their computer or something. I thought it would be fun to experiment with building a save-game editor in WebAssembly and Blazor. In theory this is an ideal framework for save-game editors, as it runs within the browser sandbox without any plugins, so a non-technical user can trust it not to mess up their computer, but it can still efficiently process binary files (and doesn't have to be written in JavaScript). The save-game file would not leave the user's computer, and the only access needed is to the file they want to edit.

I happened to be playing D2 Resurrected lately, and came across this library. I thought it might be ideal for a browser-based WASM implementation of a D2 save-game editor. Except when I looked closer, I realized that it used Reflection, which isn't supported in WASM.

So I set about rewriting the parts that used Reflection to not do that, and while I was in there, I noticed a lot of byte-array allocations that could be done away with using a mixture of `stackalloc` and `Span<byte>` with some array-pooling in certain places. I started changing those too, and I got a little carried away, upgrading to .NET 6 and adding a few other things, ending up with what you see here.

As far as I can tell, this code behaves exactly the same as the original, but a great deal faster and using much less memory. However, I did have to change the public API somewhat, so if you decide to merge this code, it's probably a good idea to increment the major version number.

Anyway, I hope you like it.

Joel