@dschu012 2021-12-31
wow, I'll try to take a closer look at it in the next day or two. appreciate all the work you put into it.