@jtmueller 2022-01-04
BTW, I've verified that this version of D2SLib can indeed run in Blazor/WASM in the browser and successfully parse save files. So it's entirely possible to make a save editor that you don't have to trust because it runs in the browser. The only thing is that because it's running in the browser, the modified save file has to be presented as a file download, which you'll have to manually move to the right folder from Downloads unless you set your browser to prompt where to save files.